+++ javascripts/base.js
@@ -166,21 +166,28 @@ Ext.define("Netzke.classes.Core.Mixin", {
         else if ((this.actions || {})[ref]) cfg = {netzkeAction: ref};
       }
 
-      if (cfg.netzkeAction) { // replace with action instance
+      if (cfg.netzkeAction) {
+        // replace with action instance
         actName = cfg.netzkeAction.camelize(true);
         if (!this.actions[actName]) throw "Netzke: unknown action " + cfg.netzkeAction;
         items[i] = this.actions[actName];
         delete(item);
-      } else if (cfg.netzkeComponent) { // replace with component config
+
+      } else if (cfg.netzkeComponent) {
+        // replace with component config
         cmpName = cfg.netzkeComponent;
         cmpCfg = this.netzkeComponents[cmpName.camelize(true)];
         if (!cmpCfg) throw "Netzke: unknown component " + cmpName;
         items[i] = Ext.apply(cmpCfg, cfg);
         delete(item);
+
       } else if (Ext.isString(cfg) && Ext.isFunction(this[cfg.camelize(true)+"Config"])) { // replace with config referred to on the Ruby side as a symbol
+        // pre-built config
         items[i] = Ext.apply(this[cfg.camelize(true)+"Config"](this.passedConfig), {netzkeParent: this});
+
       } else {
-        for (key in cfg) { // recursion
+        // recursion
+        for (key in cfg) {
           if (Ext.isArray(cfg[key])) {
             this.netzkeNormalizeConfigArray(cfg[key]);
           }