+++ test/core_test_app/app/components/simple_tab_panel.rb
@@ -5,10 +5,13 @@ class SimpleTabPanel < Netzke::Base
   end
 
   component :server_caller
+  component :hello_world do |c|
+    c.excluded = true
+  end
   component :extended_server_caller
 
   def configure(c)
-    c.items = [:server_caller, :extended_server_caller]
+    c.items = [:server_caller, :hello_world, :extended_server_caller]
     super
   end
 end